% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assmis.R
\name{assmis}
\alias{assmis}
\title{Assign NA to missing data codes.}
\usage{
assmis(dat, var, codes)
}
\arguments{
\item{dat}{data.table object.}

\item{var}{List of variable names.}

\item{codes}{List of missing data codes.}
}
\value{
Returns date.table with transformed missing data codes.
}
\description{
Assign NA to missing data codes.
}
\examples{
dat <- assmis(dat, list(dvar), list(c(9,99)))
}

