% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{lci}
\alias{uci}
\title{Computes confidence intervals (function for tables)}
\usage{
uci(x, cl = 0.95)

lci(x, cl = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{cl}{Level of confidence,}
}
\description{
Upper and lower limit of a confidence interval.
}
\examples{
x <- runif(100, 0)
uci(x)
lci(x)
}

