% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formulas.R
\name{formulas}
\alias{formulas}
\title{Formulas for linear models.}
\usage{
formulas(var1, var2, sym1, sym2)
}
\arguments{
\item{var1}{Dependent variable (character).}

\item{var2}{Independent variables (character).}

\item{sym1}{First operator (between dependent and independent variables)}

\item{sym2}{Second operator (between independent variables)}
}
\value{
Formula object.
}
\description{
Formulas for linear models.
}
\examples{
# formulas(c("var1", "var2", "~", "+"))
}

