% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitpost.R
\name{KnitPost}
\alias{KnitPost}
\title{Create posts from Rmd files for blogging.}
\usage{
KnitPost(input, output = NULL, base.url = "", fig.path = NULL)
}
\arguments{
\item{input}{Path of Rmd file.}
}
\description{
Create posts from .Rmd files for blogging.
}

