% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lrecode.R
\name{lrecode}
\alias{lrecode}
\title{Recode variables (create new variables).}
\usage{
lrecode(dat, var, newvar, old, new)
}
\arguments{
\item{dat}{Datatable object.}

\item{var}{Character or vector of characters with variable names.}

\item{newvar}{Character or vector of characters with new variable names.}

\item{old}{Old values.}

\item{new}{New values (1 to 1 recoding).}
}
\value{
Data frame with recoded variables.
}
\description{
Recode variables.
}
\details{
Avoid define the first value as missing, put missing values at the end of the old/new specification
}
\examples{
lrecode(dat, c("var1", "var2"), c("nvar1", "nvar2"), c(1,2,3,4), c(4,3,2,1))
}

