% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{Create missing data indicators}
\usage{
mi(dat, vars)
}
\arguments{
\item{dat}{`data.table` object.}

\item{vars}{Character vector with name of variables.}
}
\value{
`data.table` with indicator variables. Missing data are replaced by the mean (numeric and interger) and the mode (factors).
}
\description{
Create scores: mean or if any of the variables have specific values.
}
\examples{
a <- c(NA,2,3,4, NA)
b <- c(2,3,4,NA, 3)
c <- c(1,2,3,NA, 1)
dat <- data.table(a,b,c)
vars <- names(dat)
mi(dat, vars)
}
