% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mi_indicator.R
\name{mi_indicator}
\alias{mi_indicator}
\title{Create missing data indicators}
\usage{
mi_indicator(dat, vars = NULL)
}
\arguments{
\item{dat}{`data.table` object.}

\item{vars}{Character vector with name of variables.}
}
\value{
`data.table` with indicator variables. Missing data are replaced by the mean (numeric and interger) and the mode (factors).
}
\description{
Create missing data indicators for varaibles with missing data.
}
\examples{
a <- c(NA,2,3,4, NA)
b <- c(2,3,4,NA, 3)
c <- c(1,2,3,NA, 1)
dat <- data.table(a,b,c)
vars <- names(dat)
mi_indicator(dat, vars)
}

