% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{operations}
\alias{operations}
\alias{Mean}
\alias{Sd}
\alias{Sum}
\alias{Sums}
\alias{Max}
\alias{Min}
\title{Arithmetic and statistiscal functions but dealing with missing data.}
\usage{
Mean(x)

Sd(x)

Sum(x)

Sums(x)

Max(x)

Min(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Arithmetic and statistiscal functions but dealing with missing data.
}
\examples{
Mean(c(1,2,4, NA))
}
