% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Identify outliers.}
\usage{
outliers(x, c = 1.5, values = FALSE)
}
\arguments{
\item{c}{Thresold value.}

\item{values}{Logical. If it is \code{TRUE}, will show outlier values.}
}
\description{
Function to identify outliers
}
\examples{
outliers(x = c(1,2,3,4,100,21000,20021, -122332), values = TRUE)
}

