% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revscale.R
\name{revscale}
\alias{revscale}
\title{Reverse escale.}
\usage{
revscale(dat, var, nvar)
}
\arguments{
\item{dat}{Data.table or data.frame object.}

\item{var}{Character or vector of character with variables names to transform.}

\item{newvar}{Character or vector of character with the name of new variables.}
}
\value{
Data.table object with transformed variables.
}
\description{
Reverse escale.
}
\examples{
revscale(dat, "var", "newvar")
}

