% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowscore.R
\name{rowscore}
\alias{rowscore}
\title{Create scores}
\usage{
rowscore(dat, vars, type = "mean", p = 0.5, nitems = NULL, val = NULL,
  complete.parcel = TRUE)
}
\arguments{
\item{dat}{`data.table` object.}

\item{vars}{Character vector with name of variables.}

\item{type}{Type of score (`mean`, `any`, `parcel`). `parcel` selects the items with less missing data based on `p` (once the items are defined a mean is computed using complete cases).}

\item{p}{Proportion of items with valid values to compute mean.}

\item{nitems}{Number of items with valid values to compute mean.}

\item{complete.parcel}{If parcels use complete data only.}

\item{values}{Values used to define `any` index.}
}
\value{
Returns a vector with scores.
}
\description{
Create scores: mean or if any of the variables have specific values.
}
\examples{
a <- c(NA,2,3,4, NA)
b <- c(2,3,4,NA, 3)
c <- c(1,2,3,NA, 1)
dat <- data.table(a,b,c)
vars <- names(dat)
rowscore(dat, vars)
}

