% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savepdf.R
\name{savepdf}
\alias{savepdf}
\title{Save plots.}
\usage{
savepdf(file, width = 16, height = 10)
}
\arguments{
\item{file}{Name of the pdf file to be exported.}
}
\description{
Save plots.
}
\examples{
savepdf("myplot")
plot(x)
dev.off()
}
