% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-sdazar.R
\docType{package}
\name{sdazar}
\alias{sdazar}
\alias{sdazar-package}
\title{sdazar: My functions.}
\description{
My functions.
}
\details{
\itemize{
  \item Version: 0.2
  \item Date: 2015-04-02
  \item Depends: data.table (>= 1.9.5)
  \item URL: \url{http://github.com/sdaza/sdazar}
}
}
\author{
Sebastian Daza \email{sdaza@ssc.wisc.edu}
}

