% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{factors}
\alias{numerics}
\alias{transform}
\title{Wrapped functions for transformation of variables.}
\usage{
factors(dat, vars)

numerics(dat, vars)
}
\arguments{
\item{dat}{Data.table or data.frame object.}

\item{vars}{Character or vector of characters with variables names to transform.}
}
\value{
Data.table object with transformed variables.
}
\description{
Wrapped functions for transformation of variables.
}
\examples{
factors(dat, c("var1", "var2"))
}

