% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Computes confidence interval (CI)}
\usage{
ci(x, type = "upper", cl = 0.95)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{`upper` or `lower` limit of the confidence interval (default = `upper`).}

\item{cl}{Confidence level (default = 0.95)}
}
\value{
Returns one of the limits of the confidence interval.
}
\description{
Computes upper and lower confidence intervals of a variable
}
\examples{
x <- runif(100, 0)
ci(x)
ci(x, type = "lower")
}

