% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countmis.R
\name{countmis}
\alias{countmis}
\title{Count missing data.}
\usage{
countmis(dat, pct = TRUE)
}
\arguments{
\item{dat}{Data frame.}

\item{pct}{Logical. If this is \code{TRUE} will show percentage of missing data by variable.}
}
\description{
Function to count missing data.
}
\examples{
countmis(dat, pct = FALSE)
}

