% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitpost.R
\name{KnitPost}
\alias{KnitPost}
\title{Create posts from Rmd files for blogging.}
\usage{
KnitPost(input, base.url = "/")
}
\arguments{
\item{input}{Path of Rmd file.}
}
\description{
Create posts from .Rmd files for blogging.
}

