% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscore.R
\name{rowscore}
\alias{rowscore}
\title{Create scores}
\usage{
rowscore(dat, vars, type = "mean", p = 1/2, nitems = NULL, val = NULL)
}
\arguments{
\item{dat}{\code{data.table} or \code{data.frame}.}

\item{vars}{Character vector with name of variables.}

\item{type}{Type of score (\code{mean}, \code{any}, \code{parcel}). \code{parcel} selects items based on \code{p}. Once the items are selected the average with the available items is computed.}

\item{p}{Proportion of items with valid values to compute the average.}

\item{nitems}{Number of items with valid values to compute the average.}

\item{values}{Values used to define \code{any} index}
}
\value{
Returns a vector with scores.
}
\description{
Create scores: mean or if any of the variables have specific values.
}
\examples{
a <- c(NA,2,3,4, NA)
b <- c(2,3,4,NA, 3)
c <- c(1,2,3,NA, 1)
dat <- data.table(a,b,c)
vars <- names(dat)
rowscore(dat, vars)
}
